.claude/worktrees/happy-brown/services/requirements.txt
# --- Core FastAPI Stack ---
fastapi==0.115.4
uvicorn==0.32.0
pydantic==2.9.2
python-multipart==0.0.9
python-dotenv==1.0.1
requests==2.32.3
starlette==0.41.2
orjson==3.10.7
typing-extensions==4.12.2

# --- Google Gemini SDK ---
google-genai==0.2.1

# --- Logging & Utilities ---
loguru==0.7.2

# --- Database (MongoDB + GridFS) ---
pymongo==4.15.4
dnspython==2.6.1

# --- Cloudinary (for media uploads) ---
cloudinary==1.44.1

# --- PDF Generation & Reports ---
reportlab==4.0.4

# --- Math / Plotting ---
matplotlib==3.10.7
numpy==2.3.5

# --- Optional: Image processing safety ---
Pillow==10.4.0

# --- Document ingestion (Upload Material) ---
PyPDF2>=3.0.0,<4.0.0
python-docx>=1.1.0,<1.3.0

# --- NLP reduction pipeline (stopwords/stemming/lemmatization) ---
nltk>=3.9,<4.0
